refactor(commit): attach-diff.tsとしてdiff操作処理を分離
- commitメッセージへのdiff追加・削除処理をattach-diff.tsへ移動
- konoka-editor.tsから関連ロジックを削除し、importに変更
- コードの責務分離と可読性向上のためのリファクタリング
