+++ plugins/commit/src/konoka-editor.ts
@@ -2,6 +2,7 @@ import { Command, FileSystem } from "@effect/platform";
 import type { CommandExecutor } from "@effect/platform/CommandExecutor";
 import type { PlatformError } from "@effect/platform/Error";
 import { Config, Data, Effect } from "effect";
+import { appendDiffToEditmsg, removeDiffFromEditmsg } from "./attach-diff";
 
 export class EditorFailedError extends Data.TaggedError("EditorFailedError")<{
   readonly editor: Command.Command;
@@ -10,51 +11,6 @@ export class EditorFailedError extends Data.TaggedError("EditorFailedError")<{
   override get message(): string {
     return `Editor command "${String(this.editor)}" failed with exit code ${this.exitCode}.`;
   }
-}
-
-/**
- * Gitが`--verbose`などで、
- * コミットメッセージとdiffの仕切りに使う、
- * scissors line。
- * 各種エディタやツールはこの行を目印にして挙動を変えます。
- * scissors line以外の用途ではこのテキストは入っていないことを前提にします。
- */
-const scissorsLine = "# ------------------------ >8 ------------------------\n" as const;
-
-/**
- * `COMMIT_EDITMSG`ファイルに`git commit --verbose`のようにdiff patchデータを追加します。
- * 人間がdiffを見ながらコミットメッセージを書けるようになります。
- * またGitHub Copilotなどのツールも文脈を読み取って提案が向上します。
- */
-function appendDiffToEditmsg(
-  commitEditmsgPath: string,
-  patchPath: string,
-): Effect.Effect<void, PlatformError, FileSystem.FileSystem> {
-  return Effect.gen(function* () {
-    const fs = yield* FileSystem.FileSystem;
-    const commitEditmsg = (yield* fs.readFileString(commitEditmsgPath)).trim();
-    const patch = (yield* fs.readFileString(patchPath)).trim();
-    const content = `${commitEditmsg}\n${scissorsLine}${patch}`;
-    yield* fs.writeFileString(commitEditmsgPath, content);
-  });
-}
-
-/**
- * `COMMIT_EDITMSG`ファイルにからdiff patchデータを削除します。
- * `git commit --verbose --cleanup=scissors`で削除する方針は、
- * `commit-msg`フックなどが削除する前に参照するようになっているため、
- * scissorsを考慮しないツールでエラーを引き起こしてしまいます。
- * よってこちらでテキストエディタの編集終了後に明示的に削除します。
- */
-function removeDiffFromEditmsg(
-  commitEditmsgPath: string,
-): Effect.Effect<void, PlatformError, FileSystem.FileSystem> {
-  return Effect.gen(function* () {
-    const fs = yield* FileSystem.FileSystem;
-    const oldCommitEditmsg = yield* fs.readFileString(commitEditmsgPath);
-    const newCommitEditmsg = oldCommitEditmsg.replace(new RegExp(`${scissorsLine}.*`), "");
-    yield* fs.writeFileString(commitEditmsgPath, newCommitEditmsg);
-  });
 }
 
 /**
